# PoliTechAPI/openapi.generated.yaml
openapi: 3.0.3
info:
  title: PoliTech API (generated from controllers)
  version: 1.0.0
  description: >
    OpenAPI specification built from Spring controllers.
    WARNING: Some endpoints return placeholder/null responses in code.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/{tenantCode}/admin/clients:
    post:
      tags: [admin-clients]
      summary: Create client
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Client"
      responses:
        "200":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
      security:
        - bearerAuth: []
    get:
      tags: [admin-clients]
      summary: List clients
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
      security:
        - bearerAuth: []
  /api/v1/{tenantCode}/admin/clients/{clientId}:
    get:
      tags: [admin-clients]
      summary: Get client by id
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema:
            type: string
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
      security:
        - bearerAuth: []
    put:
      tags: [admin-clients]
      summary: Update client
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema:
            type: string
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Client"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
      security:
        - bearerAuth: []
  /api/v1/{tenantCode}/admin/clients/{clientId}/products:
    get:
      tags: [admin-clients]
      summary: List client products
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema:
            type: string
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Product roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductRole"
      security:
        - bearerAuth: []
    post:
      tags: [admin-clients]
      summary: Grant client product
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema:
            type: string
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRole"
      responses:
        "201":
          description: Created
      security:
        - bearerAuth: []
  /api/v1/{tenantCode}/admin/clients/{clientId}/products/{productRoleId}:
    delete:
      tags: [admin-clients]
      summary: Revoke client product
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema:
            type: string
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: productRoleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "202":
          description: Accepted
      security:
        - bearerAuth: []
  /api/v1/{tenantCode}/admin/accounts:
    post:
      tags: [admin-accounts]
      summary: Create account
      # WARNING: implementation currently returns a placeholder response.
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Created
      security:
        - bearerAuth: []
  /api/v1/{tenantCode}/admin/accounts/{accountId}:
    get:
      tags: [admin-accounts]
      summary: Get account
      # WARNING: implementation currently returns null.
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Account payload
          content:
            application/json:
              schema:
                type: object
      security:
        - bearerAuth: []
  /api/v1/{tenantCode}/admin/accounts/{accountId}/accounts:
    get:
      tags: [admin-accounts]
      summary: Get sub-accounts
      # WARNING: implementation currently returns null.
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Sub-accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
      security:
        - bearerAuth: []
    post:
      tags: [admin-accounts]
      summary: Create sub-account
      # WARNING: implementation currently returns a placeholder response.
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Created
      security:
        - bearerAuth: []
  /api/v1/{tenantCode}/admin/accounts/hierarchy:
    get:
      tags: [admin-accounts]
      summary: Get account hierarchy
      # WARNING: implementation currently returns null.
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hierarchy
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
      security:
        - bearerAuth: []
  /api/v1/{tenantCode}/admin/accounts/createSubaccount:
    post:
      tags: [admin-accounts]
      summary: Create subaccount (simple)
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Account"
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
      security:
        - bearerAuth: []
  /api/v1/{tenantCode}/admin/accounts/getProductRoles/{accountId}:
    get:
      tags: [admin-accounts]
      summary: Get product roles for account
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Roles
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
      security:
        - bearerAuth: []
...
  /api/admin/product-roles/account/{accountId}:
    get:
      tags: [admin-product-roles]
      summary: Get product roles by account
      # WARNING: implementation currently returns null.
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Roles
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Client:
      type: object
      x-java-class: ru.pt.api.dto.auth.Client
    ProductRole:
      type: object
      x-java-class: ru.pt.api.dto.auth.ProductRole
    CreateAccountRequest:
      type: object
      x-java-class: ru.pt.api.admin.AccountManagementController.CreateAccountRequest
    Account:
      type: object
      x-java-class: ru.pt.api.dto.auth.Account
...